Refactor LoginTokenSale
