+++ app/containers/LoginTokenSale/LoginTokenSale.js
@@ -2,14 +2,12 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { Link } from 'react-router'
-import { login } from '../modules/account'
-import { sendEvent, clearTransactionEvent } from '../modules/transactions'
-import Logo from './Logo'
-import Footer from './Footer'
-import { verifyPrivateKey } from '../core/wallet'
+import Logo from '../../components/Logo'
+import Footer from '../../components/Footer'
+import { ROUTES } from '../../core/constants'
 
 type Props = {
-  dispatch: DispatchType,
+  loginWithPrivateKey: Function,
   history: Object
 }
 
@@ -21,21 +19,12 @@ class LoginTokenSale extends Component<Props, State> {
   state = {
     wif: ''
   }
-  onWifChange = () => {
-    const { dispatch, history } = this.props
-    const { wif } = this.state
-
-    if (verifyPrivateKey(wif)) {
-      dispatch(login(wif))
-      history.push('/tokenSale')
-    } else {
-      dispatch(sendEvent(false, 'That is not a valid private key'))
-      setTimeout(() => dispatch(clearTransactionEvent()), 5000)
-    }
-  }
 
   render () {
+    const { history, loginWithPrivateKey } = this.props
     const { wif } = this.state
+    const loginButtonDisabled = wif === ''
+
     return (
       <div id='loginPage'>
         <div className='login'>
@@ -50,7 +39,10 @@ class LoginTokenSale extends Component<Props, State> {
             />
           </div>
           <div className='loginButtons'>
-            <button onClick={this.onWifChange}>Login</button>
+            <button
+              onClick={() => loginWithPrivateKey(wif, history, ROUTES.TOKEN_SALE)}
+              disabled={loginButtonDisabled}
+              className={loginButtonDisabled && 'disabled'}>Login</button>
             <Link to='/'><button className='altButton'>Home</button></Link>
           </div>
           <Footer />
