+++ app/routes.js
@@ -7,10 +7,10 @@ import TokenSale from './containers/TokenSale'
 import CreateWallet from './containers/CreateWallet'
 import Dashboard from './containers/Dashboard'
 import LoginLocalStorage from './containers/LoginLocalStorage'
+import LoginTokenSale from './containers/LoginTokenSale'
 
 import LoginNep2 from './components/LoginNep2'
 import Settings from './components/Settings'
-import LoginTokenSale from './components/LoginTokenSale'
 import EncryptKey from './components/EncryptKey'
 
 import Home from './components/Home'
